+++ Editor/CHANGES.md
@@ -1,5 +1,6 @@
 #### Version: 0.3.1
 - **Fixed**: Wrong entities position after opening same scene for the second time
+- **Removed**: Optional usage analytics 
 
 #### Version: 0.3.0
 - Old project from previous versions have to be converted using project converter before loading